Added C++ Standard
It is of utmost importance to specify a minimal standard required to build the project. Your users will both benefit and thank you for it in the long run. And, of course, you yourself will have way less hassle in ever growing projects and cross compilation on several platforms with a plethora of different compilers.